ingester/requirements.txt
psycopg2-binary
requests
redis
praw   # if Reddit API is used later
ffmpeg-python  # optional if thumbnails are generated here
